rebase to int for meal nutrients

+++ core/domain/src/main/java/com/jacqulin/calcalc/core/domain/model/Nutrition.kt
@@ -2,11 +2,11 @@ package com.jacqulin.calcalc.core.domain.model
 
 data class Nutrition(
     val name: String = "",
-    val weight: Double,
-    val calories: Double,
-    val protein: Double,
-    val fat: Double,
-    val carb: Double,
+    val weight: Int,
+    val calories: Int,
+    val protein: Int,
+    val fat: Int,
+    val carb: Int,
     val ingredient: List<Ingredient> = emptyList()
 )
 
@@ -41,7 +41,7 @@ fun Nutrition.toJsonString(): String {
                 """.trimIndent()
             )
 
-            if (index.toDouble() != ingredient.component6()) {
+            if (index != ingredient.component6()) {
                 append(",")
             }
         }
